% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{metropolis}
\alias{metropolis}
\title{Use three inputs to predict response using R.}
\usage{
metropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{the varients in normal distribution (numeric)}

\item{x0}{the first number (numeric)}

\item{N}{the number you want to get (numeric)}
}
\description{
Homework to deal with Rcpp
}
\examples{
\dontrun{
Metropolis(0.05,25,1000)
}
}
