\name{SC19088-package}
\alias{SC19088-package}
\alias{SC19088}
\docType{package}
\title{
  Final Homework
}
\description{
 Three Functions are including in SC19088.
 Bs is about comparing knots in B-spline.
 fbm is about Generating a time series of fractional Brownian motion.
 metropolis is using rcpp to do the homework in 20191220.
}
\details{
  No more details to introduct
}
\author{
SC19088, 849505139@qq.com.

Maintainer: Your Name <your@email.com>
}
\references{
  No references
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
